+++ plugin-yaml/src/main/java/net/rabbitware/config/plugin/yaml/YamlPlugin.java
@@ -116,8 +116,8 @@ public class YamlPlugin extends LocationBasedConfigSourcePlugin {
                 }
             }
             case List<?> list -> {
-                // treat uniform primitive arrays as a list
-                if (arrayIsPrimitiveAndUniform(list)) { // treat as a list
+                // treat arrays of values as a list
+                if (arrayIsAllValues(list)) { // treat as a list
                     StringBuilder sb = new StringBuilder();
                     for (int index = 0; index < list.size(); index++) {
                         if (index > 0) {
@@ -146,34 +146,15 @@ public class YamlPlugin extends LocationBasedConfigSourcePlugin {
     }
 
 
-    // returns true if the given JSONArray contains only primitive elements of
-    // the same type
-    private boolean arrayIsPrimitiveAndUniform(List<?> a) {
-        int hasString = 0;
-        int hasIntegers = 0;
-        int hasFloats = 0;
-        int hasBooleans = 0;
+    // return true if the given list contains only values
+    private boolean arrayIsAllValues(List<?> a) {
         for (int i = 0; i < a.size(); i++) {
-            switch (a.get(i)) {
-                case String s -> hasString = 1;
-                case Boolean b -> hasBooleans = 1;
-                case Number n -> {
-                    String number = n.toString();
-                    if (number.matches("[+-]?\\d+")) {
-                        hasIntegers = 1;
-                    } else if (number.matches("[+-]?(?:\\d+\\.\\d*|\\.\\d+)(?:[eE][+-]?\\d+)?|[+-]?\\d+[eE][+-]?\\d+")) {
-                        hasFloats = 1;
-                    } else {
-                        return false;
-                    }
-                }
-                case null -> hasString = 1; // treat null as a string for uniformity
-                default -> {
-                    return false;
-                }
+            Object o = a.get(i);
+            if (!(o == null || o instanceof String || o instanceof Number || o instanceof Boolean)) {
+                return false;
             }
         }
-        return hasString + hasIntegers + hasFloats + hasBooleans <= 1;
+        return true;
     }        
 
     private void add(Map<String, String> map, String key, Object value) {